test(compiler) add regression test for ParallelFor custom name with Collected
